feat(#163): support for action buttons
as requested in issue #163. the datetime-picker does still close itself
if you reach "the end of the selection". e.g. when selecting the minutes
in the minutes-view. this is in contrary to the original material date
picker, but in my personal opinion more useful. I think it should be
possible to customize this behaviour too (via property).