ci: remove deprecated protractor
message from `npm install`: protractor@7.0.0: We have news to share - Protractor
is deprecated and will reach end-of-life by Summer 2023. To learn more and find
out about other options please refer to this post on the Angular blog. Thank you
for using and contributing to Protractor. https://goo.gle/state-of-e2e-in-angular.